refactor data colloection
There are two significantly different ways to collect data from a
record: assembling a name out of rank fields, and getting name
from scientificName field. I did refactor the code to make it
possible to use different kind of 'engines' to collect data. So
now there is a Collector, that includes one or another engine
to process a row.