Fix #18 skip_original flag
When skip_original flag is set only taxonId is kept from the
original input. When it is not set -- all input is preserved.
It is helpful when the next step is not done by human but
by chaining crossmaps together or when next step is a script.

I am not implementing particular usecases, because there is no
obviously popular ones so far, and adding awk/sed/script can
do this job in a more flexible manner