Bug fix, removal of BOM, :shipit: v3.1.3
We did not have consistent number of fields in csv file.
I added tests for all possible combinations of options
to check consistency of row columns in all test files.

BOM character \u{feff} is added by Windows when it saves
UTF-8 files. It supposed to be the very first character
in the file. It prevented the first header to be recognized.
The character is now removed while we are collecting headers.