Fix #43 partially
The idea was to unify injestion and resolution to the same stats
format. However it requires significant changes, and it is better
to make injestion conurrent first, before changing its format.
So for this ticket we only create new stats format for resolution,
and when injestion will be ready, it will follow the same format
as well.